Styling thurs

+++ app.rb
@@ -1,7 +1,7 @@
 require "sinatra/base"
 require "sinatra/reloader"
 require "rack-flash"
-require 'date'
+require "date"
 
 require "./lib/database_connection.rb"
 require "./lib/space.rb"
@@ -32,12 +32,12 @@ class Application < Sinatra::Base
     if session[:maker_id]
       @maker = maker_repo.find(session[:maker_id])
     else session[:user_id]
-      @user = user_repo.find(session[:user_id])
-    end
+      @user = user_repo.find(session[:user_id])     end
     return erb(:index)
   end
 
   get "/makers/login" do
+    session[:user_id] = nil
     return erb(:makers_login)
   end
 
@@ -91,6 +91,7 @@ class Application < Sinatra::Base
         session[:maker_id] = maker.id
         redirect "/spaces"
       else
+        flash[:message] = "Invalid email or password. Please enter again."
         redirect "/makers/login"
       end
     end
@@ -122,6 +123,7 @@ class Application < Sinatra::Base
   end
 
   get "/users/login" do
+    session[:maker_id] = nil
     return erb(:users_login)
   end
 
@@ -141,6 +143,7 @@ class Application < Sinatra::Base
         session[:user_id] = user.id
         redirect "/spaces"
       else
+        flash[:message] = "Invalid email or password. Please enter again."
         return erb(:users_login)
       end
     end
@@ -155,10 +158,17 @@ class Application < Sinatra::Base
   get "/spaces/create" do
     repo = SpaceRepository.new
     maker_repo = MakersRepository.new
+    # Generate random inputs
+    @name = ["The Point Cottage, Loch Striven", "Llanelli Beach Sea View apartment", "Sea Thrift Beach House Tywyn", "Surfsplash beachfront Holiday Cottage, Dunbar"].sample
+    @location = ["London", "Manchester", "Bristol", "Cardiff", "Edinburgh"].sample
+    @description = ["Sea Thrift Beach House offers guests spacious and contemporary coastal living from an enviable beach front location, with panoramic sea views stretching across Cardigan Bay.", "The Point is a beautifully appointed remote holiday cottage on the banks of Loch Striven, Argyll, Scotland. The master bedroom features a sitting area and a balcony .The second bedroom has a double bed, robe, chest of drawers.", "First floor modern apartment situated on Carmarthenshire Coastal Path. 25 meters off Llanelli beach. The apartment offers  breath taking sea views  of Llanelli beach, Loughor estuary and across to the Gower peninsula."].sample
+    @date = ["2023-02-17", "2023-4-7", "2023-3-13"].sample
+
     if session[:maker_id]
       @maker = maker_repo.find(session[:maker_id])
       return erb(:spaces_creation)
     else
+      flash.now[:message] = "You have to log in to create a space."
       redirect "/makers/login"
     end
   end
@@ -166,26 +176,34 @@ class Application < Sinatra::Base
   post "/spaces/create" do
     space_repo = SpaceRepository.new
     maker_repo = MakersRepository.new
+
+    name = params[:name]
+    location = params[:location]
+    price = params[:price]
+    description = params[:description]
+    date = params[:date]
+
     if name.nil? || location.nil? || price.nil? || description.nil? || date.nil? || name == "" || location == "" || price == "" || description == "" || date == ""
-      flash.now[:message] = "Empty inputs are not acceptable. Please enter again."
-      return erb(:makers_signup)
+      flash[:message] = "Empty inputs are not acceptable. Please enter again."
+      redirect "/spaces/create"
     end
-    
+
     if session[:maker_id]
       @maker = maker_repo.find(session[:maker_id])
       new_space = Space.new
-      new_space.name = params[:name]
-      new_space.location = params[:location]
-      new_space.description = params[:description]
-      new_space.price = params[:price]
-      new_space.date = params[:date]
+      new_space.name = name
+      new_space.location = location
+      new_space.description = description
+      new_space.price = price
+      new_space.date = date
       new_space.available = true
       new_space.maker_id = session[:maker_id]
 
       space_repo.create(new_space)
-      
-      redirect('/spaces')
+
+      redirect("/spaces")
     else
+      flash.now[:message] = "You have to log in to create a space."
       redirect "/makers/login"
     end
   end
@@ -209,7 +227,8 @@ class Application < Sinatra::Base
       # @photos = image_links.sample
       return erb(:spaces)
     else
-      redirect "/makers/login"
+      flash[:message] = "You have to log in to browse spaces."
+      redirect "/users/login"
     end
   end
 
@@ -220,13 +239,16 @@ class Application < Sinatra::Base
     if session[:maker_id]
       @maker = maker_repo.find(session[:maker_id])
       @space = repo.find_by_id(params[:id])
+      @owner = maker_repo.find(@space.maker_id)
       return erb(:space_single)
     elsif session[:user_id]
       @user = user_repo.find(session[:user_id])
       @space = repo.find_by_id(params[:id])
+      @owner = maker_repo.find(@space.maker_id)
       return erb(:space_single)
     else
-      redirect "/"
+      flash[:message] = "You have to log in to browse spaces."
+      redirect "/users/login"
     end
   end
 
@@ -256,29 +278,68 @@ class Application < Sinatra::Base
     # if user has maker_id
     if session[:maker_id]
       maker_repo = MakersRepository.new
-      space_repo = SpaceRepository.new
+      booking_repo = BookingRepository.new
       @maker = maker_repo.find(session[:maker_id])
-      @spaces = space_repo.find_by_maker(session[:maker_id])
+      @bookings = booking_repo.find_by_maker_id(session[:maker_id])
       return erb(:maker_confirmation)
     else
       flash[:message] = "You have to log in as a maker to confirm bookings."
       redirect "/makers/login"
     end
   end
-  
+
   get "/makers/confirm_request/:space_id" do
     book_repo = BookingRepository.new
     space_repo = SpaceRepository.new
     @booking = book_repo.find_by_space_id(params[:space_id])
-    book_repo.delete(@booking.id)
-    space_repo.delete(params[:space_id])
+    book_repo.confirm(@booking.id, params[:space_id])
+    space_repo.update(params[:space_id], { available: false })
     flash[:message] = "You have confirmed a booking."
     redirect "/makers/confirmation"
   end
 
-  # get "/makers/delete_request/:space_id" do
-  #   book_repo = BookingRepository.new
-  #   @booking = book_repo.find_by_space_id(params[:space_id])
-  #   book_repo.delete(@booking.id)
-  # end
+  get "/makers/my-spaces" do
+    if session[:maker_id]
+      maker_repo = MakersRepository.new
+      spaces_repo = SpaceRepository.new
+      @maker = maker_repo.find(session[:maker_id])
+      @spaces = spaces_repo.find_by_maker(session[:maker_id])
+      return erb(:my_spaces)
+    else
+      flash[:message] = "You have to log in as a maker to view your spaces."
+      redirect "/makers/login"
+    end
+  end
+
+  get "/spaces/delete/:space_id" do
+    space_repo = SpaceRepository.new
+    maker_id = session[:maker_id].to_i
+    if maker_id
+      space = space_repo.find_by_id(params[:space_id])
+      if space.maker_id == maker_id
+        space_repo.delete(params[:space_id])
+        flash[:message] = "You have deleted '#{space.name}'."
+        redirect "/spaces"
+      else
+        flash[:message] = "You are not authorised to do so."
+        redirect "/spaces"
+      end
+    else
+      flash[:message] = "You have to log in as a maker to delete your spaces."
+      redirect "/makers/login"
+    end
+  end
+
+  get "/my-bookings" do
+    if session[:user_id]
+      book_repo = BookingRepository.new
+      user_repo = UsersRepository.new
+      @user = user_repo.find(session[:user_id])
+      @bookings = book_repo.find_by_user_id(session[:user_id])
+      return erb(:my_bookings)
+    else
+      flash[:message] = "You have to log in as a user to view your bookings."
+      redirect "/users/login"
+    end
+  end
 end
